# docs/api/campaign-openapi.yaml
openapi: 3.0.3
info:
  title: Campaign Service API
  version: 0.1.0
  description: |
    Campaigns, sessions, scenes, players, story arcs, events, and timelines for the Snapp VTT system.
    
    This specification describes the public HTTP contract of the campaign service as implemented in `apps/services/campaign/app.ts`.
servers:
  - url: https://localhost:3003
    description: Local development (CAMPAIGN_PORT or PORT)

paths:
  /campaigns:
    get:
      summary: List all campaigns
      tags: [campaigns]
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaigns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
    post:
      summary: Create a new campaign
      tags: [campaigns]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                summary:
                  type: string
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign:
                    $ref: '#/components/schemas/Campaign'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (requires gm)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /campaigns/{campaignId}/sessions:
    get:
      summary: List sessions for a campaign
      tags: [sessions]
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
    post:
      summary: Create a session for a campaign
      tags: [sessions]
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: '#/components/schemas/Session'
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (requires gm)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /sessions/{sessionId}/scenes:
    get:
      summary: List scenes for a session
      tags: [scenes]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of scenes
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Scene'
    post:
      summary: Create a scene within a session
      tags: [scenes]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                summary:
                  type: string
                worldId:
                  type: string
                entityIds:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Scene created
          content:
            application/json:
              schema:
                type: object
                properties:
                  scene:
                    $ref: '#/components/schemas/Scene'
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /campaigns/{campaignId}/players:
    get:
      summary: List players for a campaign
      tags: [players]
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Players in the campaign
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items:
                      type: string
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add a player to a campaign
      tags: [players]
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId]
              properties:
                playerId:
                  type: string
      responses:
        '201':
          description: Player added
          content:
            application/json:
              schema:
                type: object
                properties:
                  playerId:
                    type: string
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (requires gm)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /campaigns/{campaignId}/story-arcs:
    get:
      summary: List story arcs for a campaign
      tags: [storyArcs]
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of story arcs
          content:
            application/json:
              schema:
                type: object
                properties:
                  storyArcs:
                    type: array
                    items:
                      $ref: '#/components/schemas/StoryArc'
    post:
      summary: Create a story arc for a campaign
      tags: [storyArcs]
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                summary:
                  type: string
      responses:
        '201':
          description: Story arc created
          content:
            application/json:
              schema:
                type: object
                properties:
                  storyArc:
                    $ref: '#/components/schemas/StoryArc'
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (requires gm)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /story-arcs/{storyArcId}/events:
    get:
      summary: List events in a story arc
      tags: [events]
      parameters:
        - name: storyArcId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of event IDs in the story arc
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      type: string
        '404':
          description: Story arc not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add an event to a story arc
      tags: [events]
      security:
        - bearerAuth: []
      parameters:
        - name: storyArcId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId]
              properties:
                eventId:
                  type: string
      responses:
        '201':
          description: Event added to story arc
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventId:
                    type: string
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (requires gm)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /campaigns/{campaignId}/timeline:
    get:
      summary: Get the current campaign timeline
      tags: [timeline]
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Timeline information for the campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timeline'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /campaigns/{campaignId}/timeline/advance:
    post:
      summary: Advance the campaign timeline
      tags: [timeline]
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                  description: Amount to advance (default 1)
                unit:
                  type: string
                  description: Time unit
                  enum: [second, minute, hour, day, week, month, year]
      responses:
        '200':
          description: Updated timeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timeline'
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (requires gm)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Campaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        summary:
          type: string
      required: [id, name, summary]

    Session:
      type: object
      properties:
        id:
          type: string
        campaignId:
          type: string
        name:
          type: string
      required: [id, campaignId, name]

    Scene:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        name:
          type: string
        summary:
          type: string
        worldId:
          type: string
        entityIds:
          type: array
          items:
            type: string
      required: [id, sessionId, name, summary]

    StoryArc:
      type: object
      properties:
        id:
          type: string
        campaignId:
          type: string
        name:
          type: string
        summary:
          type: string
      required: [id, campaignId, name, summary]

    Timeline:
      type: object
      description: Timeline representation for a campaign.
      properties:
        currentMoment:
          type: integer
          format: int64
        unit:
          type: string
      required: [currentMoment]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]